% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_medpc.R
\name{import_medpc}
\alias{import_medpc}
\title{import data from MedPC file to an R list}
\usage{
import_medpc(filename)
}
\arguments{
\item{filename}{full path to MedPC data file}
}
\value{
list containing each variable in MedPC data file, referenced by variable name
}
\description{
import data from MedPC file to an R list
}
\examples{
example_file = system.file("extdata", "example_data", package = "rmedpc")
import_medpc(example_file)
}
